requirement.txt
langchain
langchain-openai
langchain-pinecone
langchain-community
pinecone-client
openai
python-dotenv
pymupdf